move is_witching_hour to utils module
